+++ src/main/java/Automata/WordAutomaton.java
@@ -22,43 +22,44 @@ public class WordAutomaton {
    * To be used only when this A is a DFAO (word).
    */
   public static void compareWordAutomaton(
-      Automaton A, int o, RelationalOperator.Ops operator, boolean print, String prefix, StringBuilder log) {
+      Automaton wordAutomaton, int o, RelationalOperator.Ops operator, boolean print, String prefix, StringBuilder log) {
       String opStr = operator.getSymbol();
       long timeBefore = System.currentTimeMillis();
-      UtilityMethods.logMessage(print, prefix + "comparing (" + opStr + ") against " + o + ":" + A.fa.getQ() + " states", log);
-      for (int p = 0; p < A.fa.getQ(); p++) {
-          A.fa.setOutputIfEqual(p, RelationalOperator.compare(operator, A.fa.getO().getInt(p), o));
+      UtilityMethods.logMessage(print, prefix + "comparing (" + opStr + ") against " + o + ":" + wordAutomaton.fa.getQ() + " states", log);
+      for (int p = 0; p < wordAutomaton.fa.getQ(); p++) {
+          wordAutomaton.fa.setOutputIfEqual(p, RelationalOperator.compare(operator, wordAutomaton.fa.getO().getInt(p), o));
       }
-      // As of now, A is *not* a word automaton
-      A.determinizeAndMinimize(print, prefix + " ", log);
+      // As of now, this is *not* a word automaton
+      wordAutomaton.determinizeAndMinimize(print, prefix + " ", log);
       long timeAfter = System.currentTimeMillis();
-      UtilityMethods.logMessage(print, prefix + "compared (" + opStr + ") against " + o + ":" + A.fa.getQ() + " states - " + (timeAfter - timeBefore) + "ms", log);
+      UtilityMethods.logMessage(print, prefix + "compared (" + opStr + ") against " + o + ":" + wordAutomaton.fa.getQ() + " states - " + (timeAfter - timeBefore) + "ms", log);
   }
 
   /**
    * Reverse a DFAO. Use Theorem 4.3.3 from Allouche & Shallit.
    */
-  public static void reverseWithOutput(Automaton A, boolean reverseMsd, boolean print, String prefix, StringBuilder log) {
-      if (A.fa.isTRUE_FALSE_AUTOMATON()) {
+  public static void reverseWithOutput(Automaton wordAutomaton, boolean reverseMsd,
+                                       boolean print, String prefix, StringBuilder log) {
+      if (wordAutomaton.fa.isTRUE_FALSE_AUTOMATON()) {
           return;
       }
 
       long timeBefore = System.currentTimeMillis();
-      UtilityMethods.logMessage(print, prefix + "reversing: " + A.fa.getQ() + " states", log);
+      UtilityMethods.logMessage(print, prefix + "reversing: " + wordAutomaton.fa.getQ() + " states", log);
 
-      boolean addedDeadState = A.fa.addDistinguishedDeadState(print, prefix, log);
+      boolean addedDeadState = wordAutomaton.fa.addDistinguishedDeadState(print, prefix, log);
 
       int minOutput = 0;
       if (addedDeadState) {
           // get state with smallest output. all states with this output will be removed.
           // after transducing, all states with this minimum output will be removed.
-          minOutput = A.fa.determineMinOutput();
+          minOutput = wordAutomaton.fa.determineMinOutput();
       }
 
       // need to define states, an initial state, transitions, and outputs.
       Map<Integer, Integer> newInitState = new HashMap<>();
-      for (int i = 0; i < A.fa.getQ(); i++) {
-          newInitState.put(i, A.fa.getO().getInt(i));
+      for (int i = 0; i < wordAutomaton.fa.getQ(); i++) {
+          newInitState.put(i, wordAutomaton.fa.getO().getInt(i));
       }
 
       IntList newO = new IntArrayList();
@@ -77,18 +78,18 @@ public class WordAutomaton {
           Map<Integer, Integer> currState = newStatesQueue.remove();
 
           // set up the output of this state to be g(q0), where g = currState.
-          newO.add((int) currState.get(A.fa.getQ0()));
+          newO.add((int) currState.get(wordAutomaton.fa.getQ0()));
 
           newD.add(new Int2ObjectRBTreeMap<>());
 
-          if (A.fa.getNfaD().get(A.fa.getQ0()).keySet().size() != A.getAlphabetSize()) {
+          if (wordAutomaton.fa.getNfaD().get(wordAutomaton.fa.getQ0()).keySet().size() != wordAutomaton.getAlphabetSize()) {
               throw new WalnutException("Automaton should be deterministic!");
           }
-          for (int l : A.fa.getNfaD().get(A.fa.getQ0()).keySet()) {
+          for (int l : wordAutomaton.fa.getNfaD().get(wordAutomaton.fa.getQ0()).keySet()) {
               Map<Integer, Integer> toState = new HashMap<>();
 
-              for (int i = 0; i < A.fa.getQ(); i++) {
-                  toState.put(i, currState.get(A.fa.getNfaD().get(i).get(l).getInt(0)));
+              for (int i = 0; i < wordAutomaton.fa.getQ(); i++) {
+                  toState.put(i, currState.get(wordAutomaton.fa.getNfaD().get(i).get(l).getInt(0)));
               }
 
               if (!newStatesHash.containsKey(toState)) {
@@ -104,33 +105,32 @@ public class WordAutomaton {
           }
       }
 
-      A.fa.setFields(newStates.size(), newO, newD);
+      wordAutomaton.fa.setFields(newStates.size(), newO, newD);
 
       if (reverseMsd) {
-          NumberSystem.flipNS(A.getNS());
+          NumberSystem.flipNS(wordAutomaton.getNS());
       }
 
-      minimizeSelfWithOutput(A, print, prefix + " ", log);
+      minimizeSelfWithOutput(wordAutomaton, print, prefix + " ", log);
 
       if (addedDeadState) {
-          AutomatonLogicalOps.removeStatesWithMinOutput(A, minOutput);
+          AutomatonLogicalOps.removeStatesWithMinOutput(wordAutomaton, minOutput);
       }
 
       long timeAfter = System.currentTimeMillis();
-      UtilityMethods.logMessage(print, prefix + "reversed: " + A.fa.getQ() + " states - " + (timeAfter - timeBefore) + "ms", log);
+      UtilityMethods.logMessage(print, prefix + "reversed: " + wordAutomaton.fa.getQ() + " states - " + (timeAfter - timeBefore) + "ms", log);
   }
 
   /**
-   * @param B
    * @param outputs A list of integers, indicating which uncombined automata and in what order to return.
-   * @return A list of automata, each corresponding to the list of outputs.
+   * @return A list of non-word automata, each corresponding to the list of outputs.
    * For the sake of an example, suppose that outputs is [0,1,2], then we return the list of automaton without output
    * which accepts if the output in our automaton is 0,1 or 2 respectively.
    */
-  public static List<Automaton> uncombine(Automaton B, List<Integer> outputs) {
+  public static List<Automaton> uncombine(Automaton wordAutomaton, List<Integer> outputs) {
       List<Automaton> automata = new ArrayList<>(outputs.size());
       for (Integer output : outputs) {
-          Automaton M = B.clone();
+          Automaton M = wordAutomaton.clone();
           M.fa.setOutputIfEqual(output);
           // M is *not* a word automaton
           automata.add(M);
@@ -143,10 +143,10 @@ public class WordAutomaton {
    * We minimize a DFA with output by first uncombining into automata without output, minimizing the uncombined automata, and
    * then recombining. It follows that if the uncombined automata are minimal, then the combined automata is also minimal
    */
-  public static Automaton minimizeWithOutput(Automaton B, boolean print, String prefix, StringBuilder log) {
-      IntList outputs = new IntArrayList(B.fa.getO());
+  public static Automaton minimizeWithOutput(Automaton wordAutomaton, boolean print, String prefix, StringBuilder log) {
+      IntList outputs = new IntArrayList(wordAutomaton.fa.getO());
       UtilityMethods.removeDuplicates(outputs);
-      List<Automaton> subautomata = uncombine(B, outputs);
+      List<Automaton> subautomata = uncombine(wordAutomaton, outputs);
       for (Automaton subautomaton : subautomata) {
           // These are *not* word automata
           subautomaton.determinizeAndMinimize(print, prefix, log);
@@ -158,9 +158,9 @@ public class WordAutomaton {
       return N;
   }
 
-  public static void minimizeSelfWithOutput(Automaton B, boolean print, String prefix, StringBuilder log) {
-      Automaton N = minimizeWithOutput(B, print, prefix, log);
-      B.copy(N);
+  public static void minimizeSelfWithOutput(Automaton wordAutomaton, boolean print, String prefix, StringBuilder log) {
+      Automaton N = minimizeWithOutput(wordAutomaton, print, prefix, log);
+      wordAutomaton.copy(N);
   }
 
   /**
@@ -169,18 +169,19 @@ public class WordAutomaton {
    * a DFAO that outputs o+this[x] (or this[x]+p) on input x.
    * To be used only when this automaton and M are DFAOs (words).
    */
-  public static void applyWordArithOperator(Automaton B, int o, ArithmeticOperator.Ops op, boolean reverse, boolean print, String prefix, StringBuilder log) {
+  public static void applyWordArithOperator(Automaton wordAutomaton, int o, ArithmeticOperator.Ops op, boolean reverse,
+                                            boolean print, String prefix, StringBuilder log) {
       String opStr = op.getSymbol();
       long timeBefore = System.currentTimeMillis();
-      UtilityMethods.logMessage(print, prefix + "applying operator (" + opStr + "):" + B.fa.getQ() + " states", log);
-      for (int p = 0; p < B.fa.getQ(); p++) {
-          IntList thisO = B.fa.getO();
+      UtilityMethods.logMessage(print, prefix + "applying operator (" + opStr + "):" + wordAutomaton.fa.getQ() + " states", log);
+      for (int p = 0; p < wordAutomaton.fa.getQ(); p++) {
+          IntList thisO = wordAutomaton.fa.getO();
           int thisP = thisO.getInt(p);
           thisO.set(p,
               reverse ? ArithmeticOperator.arith(op, thisP, o) : ArithmeticOperator.arith(op, o, thisP));
       }
-      minimizeSelfWithOutput(B, print, prefix + " ", log);
+      minimizeSelfWithOutput(wordAutomaton, print, prefix + " ", log);
       long timeAfter = System.currentTimeMillis();
-      UtilityMethods.logMessage(print, prefix + "applied operator (" + opStr + "):" + B.fa.getQ() + " states - " + (timeAfter - timeBefore) + "ms", log);
+      UtilityMethods.logMessage(print, prefix + "applied operator (" + opStr + "):" + wordAutomaton.fa.getQ() + " states - " + (timeAfter - timeBefore) + "ms", log);
   }
 }